feat(hotspoteditormodal): small fix improved proptypes and comments
