+++ src/components/ImageCard/Hotspot.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { Tooltip } from 'carbon-components-react';
 import { g10 } from '@carbon/themes';
@@ -7,63 +6,14 @@ import withSize from 'react-sizeme';
 
 import { settings } from '../../constants/Settings';
 import { hexToRgb } from '../../utils/componentUtilityFunctions';
+import { HotspotPropTypes } from '../../constants/SharedPropTypes';
 
-import { HotspotContentPropTypes } from './HotspotContent';
 import CardIcon from './CardIcon';
 
 const { iotPrefix } = settings;
 const { ui01, text01, ui03 } = g10;
 
-export const propTypes = {
-  /** percentage from the left of the image to show this hotspot */
-  x: PropTypes.number.isRequired,
-  /** percentage from the top of the image to show this hotspot */
-  y: PropTypes.number.isRequired,
-  /** the content of the hotspot, either a react element or an object to use the default hotspot */
-  content: PropTypes.oneOfType([
-    PropTypes.element,
-    PropTypes.shape(HotspotContentPropTypes),
-  ]).isRequired,
-  /** points to one of our enumerated icon names (ex. caretUp, edit, close)
-   * TODO: add support for the carbon icon object (svgData, viewBox, width, height)
-   */
-  icon: PropTypes.string,
-  iconDescription: PropTypes.string,
-  /** color of the hotspot */
-  color: PropTypes.string,
-  /** width of the hotspot */
-  width: PropTypes.number,
-  /** height of the hotspot */
-  height: PropTypes.number,
-  /** optional function to provide icon based on name */
-  renderIconByName: PropTypes.func,
-  /**
-   * onClick callback for when the hotspot is clicked. Returns the event and an
-   * object width the x and y coordinates */
-  onClick: PropTypes.func,
-  /** shows a border with padding when set to true */
-  isSelected: PropTypes.bool,
-  /** determines the type of hotspot to render. Defaults to 'fixed'. */
-  type: PropTypes.oneOf(['fixed', 'dynamic', 'text']),
-  /** For text hotspots, true if title should be bold */
-  bold: PropTypes.bool,
-  /** For text hotspots, true if title should be italic */
-  italic: PropTypes.bool,
-  /** For text hotspots, true if title should be underline */
-  underline: PropTypes.bool,
-  /** For text hotspots, the hexdec color of the title font, e.g. #ff0000  */
-  fontColor: PropTypes.string,
-  /** For text hotspots, the size in px of the font, e.g. 12  */
-  fontSize: PropTypes.number,
-  /** For text hotspots, the hexdec color of the background, e.g. #ff0000  */
-  backgroundColor: PropTypes.string,
-  /** For text hotspots, the opactity of the background in percentage, e.g. 100 */
-  backgroundOpacity: PropTypes.number,
-  /** For text hotspots, the hexdec color of the border, e.g. #ff0000  */
-  borderColor: PropTypes.string,
-  /** For text hotspots, the border width in px, e.g. 12  */
-  borderWidth: PropTypes.number,
-};
+const propTypes = HotspotPropTypes;
 
 const defaultProps = {
   icon: null,
