+++ src/components/ImageCard/HotspotContent.jsx
@@ -8,7 +8,6 @@
  * Office.
  */
 import React, { useRef, useState } from 'react';
-import PropTypes from 'prop-types';
 import isNil from 'lodash/isNil';
 import isEmpty from 'lodash/isEmpty';
 import { Edit20 } from '@carbon/icons-react';
@@ -19,54 +18,13 @@ import {
 } from '../../utils/cardUtilityFunctions';
 import { settings } from '../../constants/Settings';
 import { TextInput } from '../TextInput';
+import { HotspotContentPropTypes } from '../../constants/SharedPropTypes';
 
 import CardIcon from './CardIcon';
 
 const { iotPrefix } = settings;
 
-export const HotspotContentPropTypes = {
-  title: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
-  description: PropTypes.string,
-  values: PropTypes.objectOf(PropTypes.any),
-  attributes: PropTypes.arrayOf(
-    PropTypes.shape({
-      dataSourceId: PropTypes.string,
-      label: PropTypes.string,
-      unit: PropTypes.string,
-      precision: PropTypes.number,
-      thresholds: PropTypes.arrayOf(
-        PropTypes.shape({
-          comparison: PropTypes.oneOf(['<', '>', '=', '<=', '>=']),
-          value: PropTypes.any,
-          icon: PropTypes.string,
-          color: PropTypes.string,
-        })
-      ),
-    })
-  ),
-  /** overall threshold that launched the hotspot */
-  hotspotThreshold: PropTypes.shape({
-    dataSourceId: PropTypes.string,
-    comparison: PropTypes.oneOf(['<', '>', '=', '<=', '>=']),
-    value: PropTypes.any,
-    icon: PropTypes.string,
-    color: PropTypes.string,
-  }),
-  /** the locale to use for formatting numeric values */
-  locale: PropTypes.string,
-  /** The placeholder text for editable title */
-  titlePlaceholderText: PropTypes.string,
-  /** The html title attribute text for the title label when editable */
-  titleEditableHintText: PropTypes.string,
-  /** ability to render icon by name */
-  renderIconByName: PropTypes.func,
-  /** when true the title can be edited by the user. */
-  isTitleEditable: PropTypes.bool,
-  /** the unique id of this component, used by input elements */
-  id: PropTypes.string,
-  /** For text hotspots, callback with current value for when the editable fields are blurred. */
-  onChange: PropTypes.func,
-};
+const propTypes = HotspotContentPropTypes;
 
 const defaultProps = {
   title: null,
@@ -274,6 +232,6 @@ const HotspotContent = ({
   );
 };
 
-HotspotContent.propTypes = HotspotContentPropTypes;
+HotspotContent.propTypes = propTypes;
 HotspotContent.defaultProps = defaultProps;
 export default HotspotContent;
